+++ apps/gym/src/components/isolated/Weight/WeighChart/WeightChartMobile/WeightChartMobile.tsx
@@ -2,6 +2,7 @@ import styles from './WeightChartMobile.module.css'
 import { WeightChartDesktopProps } from '@/types'
 import { ChartHeader } from './ChartHeader/ChartHeader'
 import { Chart, DateRangePicker } from '@/components/reusable'
+import { Button } from '@gymapp/gymui/Button'
 
 export const WeightChartMobile = ({
   filter,
@@ -16,7 +17,11 @@ export const WeightChartMobile = ({
         startingWeight={weightData[0]}
         endingWeight={weightData[weightData.length - 1]}
       />
-      <DateRangePicker filter={filter} data={data} />
+
+      <div className={styles.button_container}>
+        <DateRangePicker filter={filter} data={data} />
+        <Button.Primary>Add +</Button.Primary>
+      </div>
       <Chart labels={labels} data={weightData} isLoading={isLoading} height='200px' />
     </div>
   )
