+++ apps/gym/src/components/isolated/Weight/index.ts
@@ -1,3 +1,2 @@
 export * from './WeightList'
 export * from './WeighChart'
-export * from './RangeSelector'
