+++ apps/gym/src/components/reusable/DatePicker/DatePicker.module.css
@@ -1,6 +1,9 @@
 .container {
   width: max-content;
   position: relative;
+  display: flex;
+  align-items: center;
+  flex-direction: column;
 }
 
 .date_input {
