--- apps/gym/src/components/reusable/DateRangePicker/DateRangePicker.module.css
@@ -1,14 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: row;
-  gap: 5px;
-  margin: 0 auto 10px auto;
-  position: relative;
-  width: auto;
-}
-
-.date_picker_container {
-  display: flex;
-  flex-flow: row wrap;
-  gap: 5px
-}
