+++ apps/gym/src/components/reusable/DateRangePicker/DateRangePickerDesktop/DateRangePickerDesktop.module.css
@@ -1,0 +1,20 @@
+.container {
+  display: flex;
+  flex-direction: row;
+  gap: 5px;
+  margin: 0 auto 10px auto;
+  position: relative;
+  width: auto;
+}
+
+.date_picker_container {
+  display: flex;
+  flex-flow: row wrap;
+  gap: 5px
+}
+
+@media (width < 480px) {
+  .container {
+    display: none;
+  }
+}
