+++ apps/gym/src/components/reusable/DateRangePicker/DateRangePickerMobile/DateRangePickerMobile.module.css
@@ -1,11 +1,12 @@
-.container {
+.datepickers {
   display: flex;
   flex-direction: column;
-  width: 100%;
+  gap: 8px;
+  align-items: center;
 }
 
-.button_container {
-  display: flex;
+.header {
+  font-size: 20px;
 }
 
 @media (width > 480px) {
