+++ apps/gym/src/components/reusable/index.ts
@@ -3,4 +3,5 @@ export * from './Sidebar'
 export * from './Searchbar'
 export * from './DatePicker'
 export * from './Dropdown'
+export * from './RangeSelector'
 export * from './DateRangePicker'
