+++ packages/gymui/src/components/Modal/FullPage.module.css
@@ -7,7 +7,7 @@
   align-items: center;
   position: fixed;
   inset: 0;
-  z-index: 100000;
+  z-index: 10000000000;
   padding: 16px;
 }
 
@@ -17,6 +17,8 @@
   display: flex;
   flex-direction: column;
   border-radius: 16px;
-  z-index: 100001;
-  
+  width: 100%;
+  align-items: center;
+  padding: 16px;
+  position: relative;
 }
